rec: Add regression tests for Proxy Protocol / ECS interaction
